backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.25
psycopg[binary]==3.2.10
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9
pydantic==2.6.3
pydantic-settings==2.2.1
alembic==1.13.1
pytest==8.1.1
httpx==0.26.0
email-validator==2.1.1
